api/requirements.txt
flask
flask-session
openai
pyyaml
python-dotenv
colorama
numpy
tiktoken
faiss-cpu
requests
tqdm
scipy
scikit-learn
pandas
